requirements.txt
flask==3.0.2
flask-sqlalchemy==3.1.1
python-dotenv==1.0.1
flask-migrate==4.0.5
flask-jwt-extended==4.6.0
flask-cors==4.0.0
requests==2.31.0
pytest==8.0.2
gunicorn==21.2.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
redis==5.0.1 
psycopg2-binary